Allow doubles for RadiusInKm and store GeohashLevels constants in a dictionary

+++ client/MAVN.Service.PartnerManagement.Client/Models/Partner/GetNearPartnersByCoordinatesRequest.cs
@@ -17,6 +17,6 @@
         /// <summary>
         /// The radius you want to search in
         /// </summary>
-        public int RadiusInKm { get; set; }
+        public double RadiusInKm { get; set; }
     }
 }
